requirements.txt
django>=3.0
gunicorn
pillow
django-crispy-forms
boto3
six
django-storages
django-ses
python-slugify
django-countries
django-timezone-field
awesome-django-timezones
django-recaptcha
